% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCDIF.R
\name{Calculate1plProb}
\alias{Calculate1plProb}
\title{Calculates the item success probability under the 1PL model.}
\usage{
Calculate1plProb(thetaValue, itemParameters, logistic = TRUE)
}
\arguments{
\item{thetaValue}{A numeric value or array for the theta (ability) value(s) where the difference will be calculated}

\item{itemParameters}{A vector or column matrix containing the numeric values of item difficulties}

\item{logistic}{A logical value stating if the IRT model will use the logistic or the normal metric.}
}
\value{
probabilities A numeric matrix with the probabilities on each thetaValue for each item.
}
\description{
Calculates the item success probability under the 1PL model.
}
\author{
Victor H. Cervantes <vhcervantesb at unal.edu.co>
}
\references{
de Ayala, R. J., (2009). The theory and practice of item response theory. New York: The Guildford Press
}

