% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCDIF.R
\name{CheckDiscriminations}
\alias{CheckDiscriminations}
\title{Identifies items with nonpositive discrimination}
\usage{
CheckDiscriminations(itemParameters)
}
\arguments{
\item{itemParameters}{A vector or column matrix containing the numeric values of item difficulties}
}
\value{
message A character string used to signal items with nonpsitive discriminations
}
\description{
Identifies items with nonpositive discrimination
}
\author{
Victor H. Cervantes <vhcervantesb at unal.edu.co>
}
