% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCDIF.R
\name{CheckGuessings}
\alias{CheckGuessings}
\title{Identifies items with guessing outside [0, 1]}
\usage{
CheckGuessings(itemParameters)
}
\arguments{
\item{itemParameters}{A vector or column matrix containing the numeric values of item difficulties}
}
\value{
message A character string used to signal items iadmissible guessing parameters
}
\description{
Identifies items with guessing outside [0, 1]
}
\author{
Victor H. Cervantes <vhcervantesb at unal.edu.co>
}
