% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCDIF.R
\name{CheckUpper}
\alias{CheckUpper}
\title{Identifies items with upper asymptote outside [0, 1]}
\usage{
CheckUpper(itemParameters)
}
\arguments{
\item{itemParameters}{A vector or column matrix containing the numeric values of item difficulties}
}
\value{
message A character string used to signal items iadmissible guessing parameters
}
\description{
Identifies items with upper asymptote outside [0, 1]
}
\author{
Victor H. Cervantes <vhcervantesb at unal.edu.co>
}
