% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGNCDIF.R
\name{IprMG}
\alias{IprMG}
\title{Item parameter replication}
\usage{
IprMG(itemParameters, itemCovariances, nReplicates = 5000)
}
\arguments{
\item{itemParameters}{A list of three or more sets of item parameters.
Item parameters are assumed to be on the same scale.
Item parameters for each group should be a matrix with nrow equal to the number of items.
One set of item parameters must be called "base" and contain the item parameters to which
all other sets will be compared to; the other group item parameters must have some name but
it can be arbitrarily given.}

\item{itemCovariances}{A list containing matrices of covariance for item estimates.
See 'itemParameters' for list structure.
Each list element may be either a list of covariance matrices for each item or a single matrix of covariance of all parameters.}

\item{nReplicates}{A numeric value indicating the number of replications to perform}
}
\value{
itemParameters A list with item parameters for focal and reference groups
}
\description{
Generates a sample of item parameters assuming multivariate normality of estimates
}
\examples{
#' # Data fom Oshima, Wright and White
baseParameters <- matrix(c(0.49, -0.07, 0.19,
                           0.92,  0.21, 0.15,
                           1.26,  0.54, 0.05,
                           0.61, -0.03, 0.18,
                           1.74,  0.01, 0.12,
                           0.50,  1.96, 0.12,
                           0.96,  0.04, 0.13,
                           0.59, -0.09, 0.18,
                           0.82, -1.16, 0.17,
                           1.26,  0.02, 0.11,
                           0.82,  0.20, 0.07,
                           0.75, -0.43, 0.15,
                           1.49, -0.06, 0.09,
                           0.97, -0.34, 0.12,
                           1.49,  0.05, 0.12,
                           0.89, -0.25, 0.15,
                           1.45,  0.06, 0.07,
                           0.75,  0.31, 0.18,
                           1.43,  0.04, 0.08,
                           0.60,  0.13, 0.22,
                           0.83,  0.52, 0.09,
                           0.56, -0.96, 0.19,
                           0.67, -0.79, 0.20,
                           0.70,  0.37, 0.18,
                           1.03, -0.71, 0.14,
                           0.89, -0.19, 0.21,
                           1.23,  0.74, 0.06,
                           0.90, -0.44, 0.18,
                           1.23, -0.17, 0.12,
                           0.69, 0.53, 0.17),
                           byrow = TRUE, ncol = 3)


group1Pars <- group2Pars <- group3Pars <- group4Pars <- group5Pars <- baseParameters

group1Pars[c(4,  22), 1] <- group1Pars[c(4,  22), 1] + .4
group1Pars[c(7,  22), 2] <- group1Pars[c(7,  22), 2] + .7

group2Pars[c(13, 28), 1] <- group2Pars[c(13, 28), 1] - .4
group2Pars[c(19, 28), 2] <- group2Pars[c(19, 28), 2] - .7

group3Pars[c(4,  13), 1] <- group3Pars[c(4,  13), 1] + .4
group3Pars[c(4,  22), 2] <- group3Pars[c(4,  22), 2] - .7

group4Pars[c(7,  28), 1] <- group4Pars[c(7,  28), 1] - .4
group4Pars[c(7,  19), 2] <- group4Pars[c(7,  19), 2] + .7

itemParameters <- list(base = baseParameters,
                       group1 = group1Pars,
                       group2 = group2Pars,
                       group3 = group3Pars,
                       group4 = group4Pars,
                       group5 = group5Pars
                       )

itemCovariances <- lapply(itemParameters, AseIrt, irtModel = "3pl", sampleSize = 5000)

mgIpr <- IprMG(itemParameters = itemParameters,
               itemCovariances = itemCovariances,
               nReplicates = 100)

}
\references{
Oshima, T. C., Wright, K., & White, N. (2014). Multiple-Group Noncompensatory Differential Item Functioning in Raju’s Differential Functioning of Items and Tests. International Journal of Testing, 15, 254–273.

Oshima, T., Raju, N. & Nanda, A. (2006). A new method for assessing the statistical significance in the Differential Functioning of Items and Tests (DFIT) framework. Journal of educational measurement, 43(1), 1--17. doi:10.1111/j.1745-3984.2006.00001.x
}
\author{
Victor H. Cervantes <vhcervantesb at unal.edu.co>
}
