% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polytomousItemParameters}
\alias{polytomousItemParameters}
\title{Sets of focal and reference item parameters from Raju et al. (2009)}
\format{a list with 'focal' and 'reference' elements. Each is a \code{matrix} 1 row per item by 5 columns: item discrimination, four item step parameters.}
\source{
This data set contains the item parameters based on those found in Raju et al. (2009).
}
\usage{
data(polytomousItemParameters)
}
\description{
This data set contains the item parameters found in Raju, N., Fortmann-Johnson, K., Kim, W., Morris, S.,
Nering, M. & Oshima, T. (2009). The item parameter replication method for detecting differential functioning in the
polytomous DFIT framework. Applied psychological measurement, 33(2), 133--147.
}
\references{
Raju, N., Fortmann-Johnson, K., Kim, W., Morris, S., Nering, M. & Oshima, T. (2009). The item parameter replication method for detecting differential functioning in the polytomous DFIT framework. Applied psychological measurement, 33(2), 133--147. doi: 10.1177/01466216083319514
}

